Pillnav converted to tabs with different styles
