--- src/components/pillnavigation/pillnavigation.css
@@ -1,23 +1,0 @@
-/**
-* Pillnavigation
-*/
-.pillnavigation .button[data-selected="true"] {
-  background: var(--color--primary);
-  border-color: var(--color--primary);
-  color: var(--fgcolor--primary);
-}
-
-.pillnavigation .button:first-of-type {
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
-}
-
-.pillnavigation .button:not(:first-of-type):not(:last-of-type) {
-  border-radius: 0;
-}
-
-.pillnavigation .button:last-of-type {
-  border-top-left-radius: 0;
-  border-right-width: 1px;
-  border-bottom-left-radius: 0;
-}
