+++ src/components/tabs/TabContent.svelte
@@ -1,6 +1,6 @@
 <script>
   import { getContext } from 'svelte';
-  import { BUTTONS } from './PillNavigation.svelte';
+  import { BUTTONS } from './Tabs.svelte';
 
   const panel = {};
   const { registerPanel, selectedPanel } = getContext(BUTTONS);
