+++ src/components/tabs/TabList.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  export let className = undefined;
+</script>
+
+<div class={className}>
+  <slot />
+</div>
