+++ src/components/tabs/tabs.css
@@ -1,0 +1,51 @@
+/**
+* Tabs
+*/
+.tabs {
+  align-items: center;
+  border-bottom: 1px solid var(--color--graa4);
+  box-sizing: border-box;
+  display: flex;
+
+  -webkit-overflow-scrolling: touch;
+}
+
+.tabs .button {
+  background: var(--color--graa6);
+  border: 1px solid var(--color--graa4);
+  border-radius: 5px 5px 0 0;
+  color: var(--color--black);
+  cursor: pointer;
+  margin: 0 0 -1px 5px;
+  padding: 5px 7px;
+}
+
+.tabs .button[data-selected="true"] {
+  background: var(--color--white);
+  border-bottom-color: var(--color--white);
+}
+
+
+/**
+* Pillnavigation
+*/
+.pillnavigation .button[data-selected="true"] {
+  background: var(--color--primary);
+  border-color: var(--color--primary);
+  color: var(--fgcolor--primary);
+}
+
+.pillnavigation .button:first-of-type {
+  border-top-right-radius: 0;
+  border-bottom-right-radius: 0;
+}
+
+.pillnavigation .button:not(:first-of-type):not(:last-of-type) {
+  border-radius: 0;
+}
+
+.pillnavigation .button:last-of-type {
+  border-top-left-radius: 0;
+  border-right-width: 1px;
+  border-bottom-left-radius: 0;
+}
